requirements.txt
flask
flask-cors
deepgram-sdk
werkzeug
ffmpeg-python